+++ src/acceptance/org/smoothbuild/acceptance/lang/GenericInferenceTest.java
@@ -13,7 +13,7 @@ import org.smoothbuild.acceptance.testing.Concat;
 public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void actual_result_type_can_be_inferred_from_arguments() throws Exception {
-    givenScript("testIdentity(a value) = value;        \n"
+    givenScript("testIdentity(A value) = value;        \n"
         + "      result = testIdentity(value='abc');   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -22,7 +22,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
 
   @Test
   public void actual_result_type_can_be_inferred_from_arguments_and_converted() throws Exception {
-    givenScript("a myfunc(a res, a forcedType) = res;               \n" +
+    givenScript("A myfunc(A res, A forcedType) = res;               \n" +
         "        result = myfunc(res=[], forcedType=[['abc']]);     \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -32,19 +32,19 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void actual_array_type_can_be_inferred_from_arguments_and_elements_are_converted()
       throws Exception {
-    givenScript("pair(a first, a second) = [first, second];           \n" +
+    givenScript("pair(A first, A second) = [first, second];           \n" +
         "        result = pair(first=[], second=[['aaa']]);           \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
     then(artifactArray("result"), equalTo(list(list(), list(list("aaa")))));
   }
 
-  // testConcat([a] first, [a] second)
+  // testConcat([A] first, [A] second)
 
   @Test
   public void infer_actual_type_of_parameters_in_concat_function_0() throws Exception {
     givenNativeJar(Concat.class);
-    givenScript("[a] testConcat([a] first, [a] second);      \n"
+    givenScript("[A] testConcat([A] first, [A] second);      \n"
         + "      result = testConcat(first=[], second=[]);   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -54,7 +54,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_concat_function_1() throws Exception {
     givenNativeJar(Concat.class);
-    givenScript("[a] testConcat([a] first, [a] second);                \n"
+    givenScript("[A] testConcat([A] first, [A] second);                \n"
         + "      result = testConcat(first=['aaa'], second=['bbb']);   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -64,7 +64,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_concat_function_2() throws Exception {
     givenNativeJar(Concat.class);
-    givenScript("[a] testConcat([a] first, [a] second);           \n"
+    givenScript("[A] testConcat([A] first, [A] second);           \n"
         + "      result = testConcat(first=['aaa'], second=[]);   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -74,7 +74,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_concat_function_3() throws Exception {
     givenNativeJar(Concat.class);
-    givenScript("[a] testConcat([a] first, [a] second);           \n"
+    givenScript("[A] testConcat([A] first, [A] second);           \n"
         + "      result = testConcat(first=[], second=['bbb']);   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -84,7 +84,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_concat_function_4() throws Exception {
     givenNativeJar(Concat.class);
-    givenScript("[a] testConcat([a] first, [a] second);           \n"
+    givenScript("[A] testConcat([A] first, [A] second);           \n"
         + "      result = testConcat(first=[[]], second=['bbb']);   \n");
     whenSmoothBuild("result");
     thenFinishedWithError();
@@ -95,8 +95,8 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_concat_function_5() throws Exception {
     givenNativeJar(Concat.class);
-    givenScript("[a] testConcat([a] first, [a] second);                            \n"
-        + "      wrapper([Nothing] f, [[a]] s) = testConcat(first=f, second=s);    \n"
+    givenScript("[A] testConcat([A] first, [A] second);                            \n"
+        + "      wrapper([Nothing] f, [[A]] s) = testConcat(first=f, second=s);    \n"
         + "      result = wrapper(f=[], s=[['aaa']]);                              \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -106,21 +106,21 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_concat_function_6() throws Exception {
     givenNativeJar(Concat.class);
-    givenScript("[a] testConcat([a] first, [a] second);                             \n"
-        + "      [a] testConcatW([a] f, [a] s) = testConcat(first=f, second=s);     \n"
-        + "      wrapper([Nothing] f, [[a]] s) = testConcatW(f=f, s=s);             \n"
+    givenScript("[A] testConcat([A] first, [A] second);                             \n"
+        + "      [A] testConcatW([A] f, [A] s) = testConcat(first=f, second=s);     \n"
+        + "      wrapper([Nothing] f, [[A]] s) = testConcatW(f=f, s=s);             \n"
         + "      result = wrapper(f=[], s=[['aaa']]);                               \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
     then(artifactArray("result"), equalTo(list(list("aaa"))));
   }
 
-  // testAppend([a] array, a element)
+  // testAppend([A] array, a element)
 
   @Test
   public void infer_actual_type_of_parameters_in_append_function_0() throws Exception {
     givenNativeJar(Append.class);
-    givenScript("[a] testAppend([a] array, a element);               \n"
+    givenScript("[A] testAppend([A] array, A element);               \n"
         + "      result = testAppend(array=[], element='bbb');       \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -130,7 +130,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_append_function_1() throws Exception {
     givenNativeJar(Append.class);
-    givenScript("[a] testAppend([a] array, a element);                \n"
+    givenScript("[A] testAppend([A] array, A element);                \n"
         + "      result = testAppend(array=['aaa'], element='bbb');   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -140,7 +140,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_append_function_2() throws Exception {
     givenNativeJar(Append.class);
-    givenScript("[a] testAppend([a] array, a element);                              \n"
+    givenScript("[A] testAppend([A] array, A element);                              \n"
         + "      StringStruct { String value }                                      \n"
         + "      result = testAppend(array=['aaa'], element=stringStruct('bbb'));   \n");
     whenSmoothBuild("result");
@@ -151,7 +151,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_append_function_3() throws Exception {
     givenNativeJar(Append.class);
-    givenScript("[a] testAppend([a] array, a element);                                      \n"
+    givenScript("[A] testAppend([A] array, A element);                                      \n"
         + "      StringStruct { String value }                                              \n"
         + "      [String] emptyStringArray = [];                                            \n"
         + "      result = testAppend(array=emptyStringArray, element=stringStruct('bbb'));  \n");
@@ -163,7 +163,7 @@ public class GenericInferenceTest extends AcceptanceTestCase {
   @Test
   public void infer_actual_type_of_parameters_in_append_function_4() throws Exception {
     givenNativeJar(Append.class);
-    givenScript("[a] testAppend([a] array, a element);                              \n"
+    givenScript("[A] testAppend([A] array, A element);                              \n"
         + "      StringStruct { String value }                                      \n"
         + "      result = testAppend(array=[[]], element=stringStruct('bbb'));      \n");
     whenSmoothBuild("result");