+++ src/acceptance/org/smoothbuild/acceptance/lang/GenericTest.java
@@ -12,7 +12,7 @@ public class GenericTest extends AcceptanceTestCase {
   @Test
   public void flatten_1() throws Exception {
     givenNativeJar(Flatten.class);
-    givenScript("[e] testFlatten([[e]] array);                         \n"
+    givenScript("[E] testFlatten([[E]] array);                         \n"
         + "      result = testFlatten(array=[['aa'], ['bb', 'cc']]);   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -22,7 +22,7 @@ public class GenericTest extends AcceptanceTestCase {
   @Test
   public void flatten_sample_2() throws Exception {
     givenNativeJar(Flatten.class);
-    givenScript("[e] testFlatten([[e]] array);                           \n"
+    givenScript("[E] testFlatten([[E]] array);                           \n"
         + "      result = testFlatten(array=[[['aa'], ['bb', 'cc']]]);   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
@@ -31,9 +31,9 @@ public class GenericTest extends AcceptanceTestCase {
 
   @Test
   public void pair_and_identity() throws Exception {
-    givenScript("a testIdentity(a v) = v;                                           \n"
-        + "      [a] pair(a e1, a e2) = [e1, e2];                                   \n"
-        + "      result = pair(e1=testIdentity(v='aa'), e2=testIdentity(v='bb'));   \n");
+    givenScript("A testIdentity(A v) = v;                                           \n"
+        + "      [A] pair(A a1, A a2) = [a1, a2];                                   \n"
+        + "      result = pair(a1=testIdentity(v='aa'), a2=testIdentity(v='bb'));   \n");
     whenSmoothBuild("result");
     thenFinishedWithSuccess();
     then(artifactArray("result"), equalTo(list("aa", "bb")));